convert all Table related components from .js => .tsx
