+++ application/react/_global/Table/ExpandableRow/ExpandableRow.tsx
@@ -1,11 +1,18 @@
 import React from 'react'
 import Row from '../Row/Row'
 import './ExpandableRow.scss'
+import { RowData, Row as tRow } from '@tanstack/react-table'
 
 const ExpandableRow = (props) => {
-  const { row, children } = props
-  const [expanded, setExpanded] = React.useState(false)
+  const { 
+    row, 
+    children 
+  } : {
+    row: tRow<RowData>,
+    children: React.ReactNode
+  } = props;
 
+  const [expanded, setExpanded] = React.useState(false)
   const toggleExpanded = () => setExpanded(!expanded);
 
   return (
